No actual changes here but there are some caveats to the way WordPress handles the homepage. See http://codex.wordpress.org/Template_Hierarchy
Index.php is a fallback template which is called as a last resort. However a lot of themes, including ones I write, use index.php as the homepage. I don't really have a good idea of what would be better than the homepage…

home.php is the listing of blog posts whether that is at / or /blog for example. In the reading settings the admin can specify if the front page is a static page or a list of blog posts.

front-page.php is what is used when a static page is selected from the reading settings.

This is confusing buy play around with it to see what I mean.